refactor: 인증 관련 기능 및 폴더 구조 개선
